Start the daemon from the tray
If the daemon is in a bad state, restarting the tray will also restart the 
daemon.

The daemon watches stdin and exits if stdin is closed. This prevents the user
to run multiple daemon.

This requires to ship crc inside the tray application. It enables changing the
.pkg to deliver a single /Application/CodeReady Containers.app. 
When upgrading, we will have to ask the user to stop the tray application.